requirements.txt
absl-py==2.3.1
anndata==0.9.2
astunparse==1.6.3
cachetools==5.5.2
certifi==2025.10.5
charset-normalizer==3.4.4
contourpy==1.1.1
cycler==0.12.1
flatbuffers==25.9.23
fonttools==4.57.0
gast==0.4.0
get-annotations==0.1.2
google-auth==2.42.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.11.0
idna==3.11
igraph==0.11.8
importlib_metadata==8.5.0
importlib_resources==6.4.5
jax==0.4.13
joblib==1.4.2
keras==2.12.0
kiwisolver==1.4.7
leidenalg==0.10.2
libclang==18.1.1
llvmlite==0.41.1
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
ml-dtypes==0.2.0
natsort==8.4.0
networkx==3.1
numba==0.58.1
numpy==1.23.5
oauthlib==3.3.1
opt_einsum==3.4.0
packaging==25.0
pandas==2.0.3
patsy==1.0.2
pillow==10.4.0
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pynndescent==0.5.13
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
scanpy==1.9.8
scikit-learn==1.3.2
scikit-misc==0.2.0
scipy==1.10.1
seaborn==0.13.2
session_info==1.0.1
six==1.17.0
statsmodels==0.14.1
stdlib-list==0.10.0
tabulate==0.9.0
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.12.0
tensorflow-metal==1.0.0
termcolor==2.4.0
texttable==1.7.0
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.5.0
tzdata==2025.2
umap-learn==0.5.7
urllib3==2.2.3
Werkzeug==3.0.6
wrapt==1.14.2
zipp==3.20.2
